rename image_2d
