+++ source/consumers/CMakeLists.txt
@@ -9,7 +9,7 @@ set(SOURCES
   ${dir}/histogram_1d.cpp
   ${dir}/histogram_2d.cpp
   ${dir}/histogram_3d.cpp
-  ${dir}/image_2d.cpp
+  ${dir}/Image2D.cpp
   ${dir}/Prebinned1D.cpp
   ${dir}/Spectrum.cpp
   ${dir}/TimeSpectrum.cpp
@@ -27,7 +27,7 @@ set(HEADERS
   ${dir}/histogram_1d.h
   ${dir}/histogram_2d.h
   ${dir}/histogram_3d.h
-  ${dir}/image_2d.h
+  ${dir}/Image2D.h
   ${dir}/Prebinned1D.h
   ${dir}/Spectrum.h
   ${dir}/TimeSpectrum.h
