+++ source/consumers/Image2D.cpp
@@ -1,4 +1,4 @@
-#include <consumers/image_2d.h>
+#include <consumers/Image2D.h>
 
 #include <consumers/dataspaces/sparse_map2d.h>
 #include <consumers/dataspaces/sparse_matrix2d.h>
