+++ tests/consumers/CMakeLists.txt
@@ -4,7 +4,7 @@ set(SOURCES
   ${dir}/histogram_1d.cpp
   ${dir}/histogram_2d.cpp
   ${dir}/histogram_3d.cpp
-  ${dir}/image_2d.cpp
+  ${dir}/Image2DTest.cpp
   ${dir}/Prebinned1DTest.cpp
   ${dir}/spectrum.cpp
   ${dir}/TimeSpectrumTest.cpp
